Update all metrics values and improve logging
All metrics will be uploaded from UptimeRobot to CachetHQ. Unfortunately,
they don't share an identifier, so we have to use their datetimes to try to
avoid creating the same metric point twice. This is a bit unreliable, because
UptimeRobot and CachetHQ might be in different timezones (UptimeRobot doesn't
allow us to configure the timezone, using whatever the user picked).

It's the best we can do so far, if we want to avoid having to extract all
UptimeRobot and CachetHQ's metric points every time we update them.

Update: still make metric optional